requirements.txt
filelock==3.18.0
fsspec==2025.5.1
grpcio==1.71.0
grpcio-reflection==1.71.0
grpcio-tools==1.71.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
paho-mqtt==2.1.0
pandas==2.2.3
protobuf==5.29.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.2.2
scipy==1.15.3
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
triton==3.2.0
typing_extensions==4.13.2
tzdata==2025.2